feat: error levels
Allow configuration of error levels for rules, which can be either `LintLevel::Error` or `LintLevel::Warning`.

This is configured via the `level` key in each rule's settings, which is now a reserved key. While it is available for rule setup to read if desired, trying to use this for other purposes will conflict with the error level logic.